backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

numpy==2.0.2
pandas==2.2.3
scikit-learn==1.6.1
